fix FTBFS
AOSP Nougat turns on lzma support on libunwind, but it seems that's
going to take some time to implement it. On Debian, three lzma related
libraries available: liblzma-dev, lzma-dev and 7z.so from p7zip-full.
Android uses a fork from the last one, which comes with no header in
Debian default install. Official libunwind uses liblzma-dev, and that
has been written to prefer p7zip in Android.

So, by turning off lzma dependency an trivial error jumps out and is
fixed with this change.